requirements.txt
playwright
browser-use
python-dotenv